Fix connect
