+++ requirements.txt
@@ -1,4 +1,4 @@
-# adafruit_io==2.7.1
+adafruit_io==2.7.2
 Flask==3.0.3
 Flask_Bcrypt==1.0.1
 flask_cookies==0.0.1
